requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scikit-learn>=1.3.0
pandas>=2.0.0
matplotlib>=3.7.0
onnx>=1.14.0
onnxruntime>=1.15.0

# Phase 1.2: REST API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6
httpx>=0.25.0  # For API testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Phase 1.3: Production Monitoring
prometheus-client>=0.19.0
psutil>=5.9.0

# Phase 2.1: Security Engine Integration
yara-python>=4.2.0